bloodDonationPlatform/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-phonenumber-field==8.0.0
phonenumbers==8.13.39
sqlparse==0.5.0
typing_extensions==4.12.2
